Рефакторинг модуля создания мока для карточки фильма (кроме комментариев): - переработан ряд функций генерации данных - константы вынесены в отдельный модуль - вспомогательные функции вынесены в отдельный модуль
