+++ package.json
@@ -19,5 +19,7 @@
   "engines": {
     "node": "10"
   },
-  "dependencies": {}
+  "dependencies": {
+    "dayjs": "1.9.5"
+  }
 }
