+++ src/view/constants.js
@@ -1,0 +1,164 @@
+const TITLES = [
+  `The Dance of Life`,
+  `Sagebrush Trail`,
+  `The Man with the Golden Arm`,
+  `Santa Claus Conquers the Martians`,
+  `Popeye the Sailor Meets Sindbad the Sailor`
+];
+
+const ORIGIN_TITLES = [
+  `The Dance of Life`,
+  `Sagebrush Trail`,
+  `The Man with the Golden Arm`,
+  `Santa Claus Conquers the Martians`,
+  `Popeye the Sailor Meets Sindbad the Sailor`
+];
+
+const POSTER_FOLDER = `./images/posters/`;
+
+const POSTER_FILE_NAMES = [
+  `made-for-each-other.png`,
+  `popeye-meets-sinbad.png`,
+  `sagebrush-trail.jpg`,
+  `santa-claus-conquers-the-martians.jpg`,
+  `the-dance-of-life.jpg`,
+  `the-great-flamarion.jpg`,
+  `the-man-with-the-golden-arm.jpg`
+];
+
+const NAMES = [
+  `Anthony Mann`,
+  `Danna Kolleen`,
+  `Sheridan Edison`,
+  `Bryant Graham`,
+  `Leroi Karson`,
+  `Carlisle Sonnie`,
+  `Arlene Bryant`,
+  `Aubrie Warrick`,
+  `Sonya Audra`,
+  `Christian Imogene`,
+  `Nash Nova`,
+  `Rosalynne Sukie`,
+  `Emil Debby`,
+  `Laverne Melody`,
+  `Nic Calla`,
+  `Donald Roy`,
+  `Derby Jenni`,
+  `Shelagh Kacie`,
+  `Mackenzie Shanae`,
+  `Gillian Chuck`,
+  `Elfreda Breanne`,
+  `Abby Kelda`,
+  `Pam Warner`,
+  `Piper Gordon`,
+  `Lawrence Flick`,
+  `Annmarie Benson`,
+  `Miranda Emelia`,
+  `Bernard Ridge`,
+  `Bettie Nigella`,
+  `Orville Emily`,
+  `Jayme Jaiden`,
+  `Leigh Cyprian`,
+  `Yolanda Ollie`,
+  `Julianna Watson`,
+  `Mort Grosvenor`,
+  `Dena Joisse`,
+  `Hercules Seale`,
+  `Valry Harder`,
+  `Rabi Fraumeni`,
+  `Ellette De costa`,
+  `Cristian Biscay`,
+  `Shel Hiney`,
+  `Gerhardt Rakkhita`,
+  `Truda Hernandez`,
+  `Damian Hickson`
+];
+
+const MAX_WRITERS_COUNT = 3;
+const MAX_ACTORS_COUNT = 3;
+
+const FIRST_FILM_DATE = `12.28.1895`;
+
+const COUNTRIES = [
+  `Albania`,
+  `Bahamas`,
+  `Cambodia`,
+  `Denmark`,
+  `Ecuador`,
+  `France`,
+  `Germany`,
+  `Hungary`,
+  `India`,
+  `Japan`,
+  `Kazakhstan`,
+  `Latvia`,
+  `Malaysia`,
+  `USA`,
+  `Poland`,
+  `Russian Federation`,
+  `Uganda`
+];
+
+const MIN_DURATION = 1;
+const MAX_DURATION = 300;
+
+const GENRES = [
+  `Musicals`,
+  `Horror`,
+  `Crime`,
+  `Romantic Comedy`,
+  `Drama`,
+  `Fantasy`,
+  `Animation`,
+  `Historical`,
+  `Science fiction`,
+  `Action Adventure`
+];
+const MAX_GENRES_COUNT = 3;
+
+const MIN_RATING = 1;
+const MAX_RATING = 10;
+
+const SENTENCES = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
+  `Fusce tristique felis at fermentum pharetra.`,
+  `Aliquam id orci ut lectus varius viverra.`,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
+  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
+  `Aliquam erat volutpat.`,
+  `Nunc fermentum tortor ac porta dapibus.`,
+  `In rutrum ac purus sit amet tempus.`
+];
+const MAX_SENTENCES = 5;
+
+const AGE_RATINGS = [
+  `0+`,
+  `6+`,
+  `12+`,
+  `16+`,
+  `18+`
+];
+
+export {
+  TITLES,
+  ORIGIN_TITLES,
+  POSTER_FOLDER,
+  POSTER_FILE_NAMES,
+  NAMES,
+  MAX_WRITERS_COUNT,
+  MAX_ACTORS_COUNT,
+  FIRST_FILM_DATE,
+  COUNTRIES,
+  MIN_DURATION,
+  MAX_DURATION,
+  GENRES,
+  MAX_GENRES_COUNT,
+  MIN_RATING,
+  MAX_RATING,
+  SENTENCES,
+  MAX_SENTENCES,
+  AGE_RATINGS
+};
